Put JsProxy in it's own header file and add a CPP file.  This makes fleshing it out easier.
