+++ src/ui/ui.pro
@@ -85,6 +85,7 @@ SOURCES += main.cpp\
     clickablelabel.cpp \
     frmencodingchooser.cpp \
     EditorNS/bannerindentationdetected.cpp \
+    EditorNS/jsproxy.cpp \
     frmindentationmode.cpp \
     singleapplication.cpp \
     localcommunication.cpp \
@@ -110,7 +111,7 @@ SOURCES += main.cpp\
     keygrabber.cpp \
     Sessions/sessions.cpp \
     Sessions/persistentcache.cpp \
-    nqqsettings.cpp \  
+    nqqsettings.cpp \
     nqqrun.cpp
 
 HEADERS  += include/mainwindow.h \
@@ -126,6 +127,7 @@ HEADERS  += include/mainwindow.h \
     include/EditorNS/bannerbasicmessage.h \
     include/EditorNS/bannerfileremoved.h \
     include/EditorNS/customqwebview.h \
+    include/EditorNS/jsproxy.h \
     include/clickablelabel.h \
     include/frmencodingchooser.h \
     include/EditorNS/bannerindentationdetected.h \
